Message acknowledgement determined by core function
It was a question whether a request should be acknowledged or not. It is hard
(impossible?) to define when to acknowledge a request automatically. The
solution to this is to let the core function specify what it wants,
through the `finalize` attribute of a Response object.

The default is True; that is, the request will be acknowledged if the
core function does not specify anything.